--- src/components/tabs/Tasks/ToDoContainer.jsx
@@ -1,5 +1,0 @@
-// import React from 'react';
-
-// class ToDoContainer extends React.Component {
-
-// }
